+++ UnitTest/Execute/DisplacementMappingDemo/PNTriangleDemo.cpp
@@ -42,7 +42,7 @@ void sdt::PNTriangleDemo::Initialize()
 	IBuffer = new IndexBuffer(Indices.data(), Indices.size());
 
 	WVPBuffer_DS = new ConstantBuffer(
-		(UINT)ShaderType::DomainShader,
+		ShaderType::DomainShader,
 		0,
 		&WVP,
 		"World View Projection",
@@ -51,7 +51,7 @@ void sdt::PNTriangleDemo::Initialize()
 	);
 	
 	TessellationFactorBuffer_HS = new ConstantBuffer(
-		(UINT)ShaderType::HullShader,
+		ShaderType::HullShader,
 		1,
 		&TessellationFactor,
 		"TessellationFactor",
