+++ UnitTest/Execute/OceanDemo/TextureTransposeDemo.cpp
@@ -37,7 +37,7 @@ void sdt::TextureTransposeDemo::Initialize()
 	);
 
 	CBuffer->BindToGPU();
-	InputTexture->BindToGPU(0, static_cast<UINT>(ShaderType::ComputeShader));
+	InputTexture->BindToGPU(0,(ShaderType::ComputeShader));
 	TestOutTexture->BindToGPUAsUAV(0);
 	TestComputeShader->Dispatch(ConstData.Width / ThreadX, ConstData.Height / ThreadY,1);
 	TestOutTexture->SaveOutputAsFile(L"Transposed");
@@ -70,7 +70,7 @@ void sdt::TextureTransposeDemo::Render()
 	// 	TransposeComputeShader->Dispatch(nullptr, 1,1,1);
 	//
 	// if (!!Original)
-	// 	Original->BindToGPU(0, static_cast<UINT>(ShaderType::PixelShader));
+	// 	Original->BindToGPU(0,(ShaderType::PixelShader));
 	// if (!!Transposed)
 	// 	Transposed->BindToGPU(1);
 	// if (!!OriginalShader)
